view 에서 delegate 구현했던 것을 ViewController 로 이동
