+++ WebviewApp/WebviewApp/ContainerView.swift
@@ -8,9 +8,8 @@
 
 import UIKit
 import WebKit
-import SafariServices
 
-class ContainerView: UIView, WKNavigationDelegate, WKScriptMessageHandler {
+class ContainerView: UIView {
     private var webView: WKWebView?
     
     required init(coder aDecoder: NSCoder) {
@@ -20,7 +19,6 @@ class ContainerView: UIView, WKNavigationDelegate, WKScriptMessageHandler {
         webView = WKWebView(frame: self.bounds, configuration: config)
         if let webView = webView {
             addSubview(webView)
-            webView.navigationDelegate = self
         }
     }
     
@@ -29,36 +27,13 @@ class ContainerView: UIView, WKNavigationDelegate, WKScriptMessageHandler {
             return
         }
         let request = URLRequest(url: url)
-        webView!.load(request)
-    }
-    
-    internal func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-        guard let scriptPath = Bundle.main.path(forResource: "querySelect", ofType: "js"),
-            var scriptSource = try? String(contentsOfFile: scriptPath) else { return }
-        scriptSource.removeLast()
-        webView.evaluateJavaScript(scriptSource, completionHandler: nil)
-    }
-    
-    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
-        guard navigationAction.navigationType == .linkActivated else {
-            decisionHandler(.allow)
-            return
-        }
-        
-        if let url = navigationAction.request.url, url.path.hasSuffix("menu.nhn") {
-            safariWebViewPresent(url: url)
-            decisionHandler(.cancel)
-            return
-        }
-        
-        decisionHandler(.allow)
+        webView?.load(request)
     }
     
     private func makeWebViewConfig(javaScriptSource: String) -> WKWebViewConfiguration {
         let contentController = WKUserContentController()
         let script = WKUserScript(source: javaScriptSource, injectionTime: .atDocumentEnd, forMainFrameOnly: true)
         contentController.addUserScript(script)
-        contentController.add(self, name: "jsHandler")
         
         let config = WKWebViewConfiguration()
         config.userContentController = contentController
@@ -66,20 +41,11 @@ class ContainerView: UIView, WKNavigationDelegate, WKScriptMessageHandler {
         return config
     }
     
-    func safariWebViewPresent(url: URL) {
-        let vc = SFSafariViewController(url: url)
-        
-        if var topController = UIApplication.shared.keyWindow?.rootViewController {
-            while let presentedViewController = topController.presentedViewController {
-                topController = presentedViewController
-            }
-            
-            topController.present(vc, animated: true, completion: nil)
-        }
+    func setWebViewNavigationDelegate(viewController: ViewController) {
+        webView?.navigationDelegate = viewController
     }
     
-    func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        let menus = Menus(body: message.body)
-        dump(menus)
+    func addScriptMessageHandler(viewController: ViewController) {
+        webView?.configuration.userContentController.add(viewController, name: "jsHandler")
     }
 }
